pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ezctfer"
version = "1.0.0"
description = "基于LangChain的多Agent CTF结队解题框架"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain>=1.2.10",
    "langgraph>=1.0.10",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=14.0",
    "httpx>=0.25.0",
    "psutil>=5.9.0",
    "idapro==0.0.7",
    "langchain-deepseek>=1.0.1",
    "ida_pro_mcp==1.4.0",
    "lightrag-hku>=1.4.8",
    "torch",
    "sentence-transformers>=5.2.0",
]

[project.scripts]
ezctfer = "ezctfer.__main__:cli_entry"

[tool.hatch.build.targets.wheel]
packages = ["src/ezctfer"]

[tool.hatch.build.targets.sdist]
include = ["src/ezctfer"]

[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
